+++ kontrol/filter/optimize.py
@@ -4,14 +4,14 @@ This is an ad hoc premature filter optimization python package for designing
 and optimization of digital filters.
 """
 
-from control import tf
+from control import tf, h2syn, ss
 import numpy as np
 from scipy.optimize import (
                             dual_annealing,
                             differential_evolution,
                             minimize, Bounds)
 import time
-from ..utils import quad_sum, norm2
+from ..utils import quad_sum, norm2, tfmatrix2tf
 # from .filters import complementary_sekiguchi, complementary_modified_sekiguchi
 
 # We are using functions for filters for now, for simplicity.
@@ -122,3 +122,52 @@ def optimize_complementary_filter(complementary_filter, spectra, f, \
     print('Done. Time taken: %.2f s The 2-norm is %.2f unit'
     %(time.clock()-t0, result.fun))
     return(result)
+
+def h2complementary(n1, n2):
+    '''H2 optimal complementary filter synthesis
+
+    Parameters
+    ----------
+        n1: control.xferfcn.TransferFunction
+            The transfer function representing the noise content of a
+            particular sensor to be fused.
+        n2: control.xferfcn.TransferFunction
+            The trasnfer function representing the noise content of another
+            sensor to be fused.
+
+    Returns
+    -------
+        h1: control.xferfcn.TransferFunction
+            The complementary filter filtering n1.
+        h2: control.xferfcn.TransferFunction
+            The complementary filter filtering n2.
+
+    Notes
+    -----
+    This function ultilizes control.robust.h2syn which depends on the slycot
+    module. If you are using under a conda virtual environment, the slycot
+    module can be installed easily from conda-forge. Using pip to install
+    slycot is a bit more involved (I have yet to suceed installing slycot in
+    my Windows machine). Please refer to the python-control package for further
+    instructions.
+
+    It is possible that h2syn yields no solution for some tricky noise profiles
+    . Try adjusting the noise profiles at some irrelevant frequencies.
+
+    Thomas Dehaeze [1]_ had the idea first so credits goes to him. (Properly
+    cite when the paper is published.)
+
+    References
+    ----------
+    .. [1]
+        Dehaeze, T. https://tdehaeze.github.io/dehaeze20_optim_robus_compl_fil
+        te/matlab/index.html
+    '''
+    p = [[tf([0],[1]), n2, tf([1],[1])],
+         [n1, -n2, tf([0],[1])]]
+#     p = [[n1, -n1],
+#          [tf([0],[1]), n2],
+#          [tf([1],[1]), tf([0],[1])]]
+    p = tfmatrix2tf(p)
+    h1 = h2syn(ss(p), 1, 1)
+    return(k1)